fix(auth): extract createPayment to its own module to satisfy bundle cap
CI bundlemon failed: dist/esm/auth/signup.js was 2.67KB > 2.5KB after the
contact-info-ordering and zero-amount-preview-swallow fixes added during
review.

Moving createPayment + planNameFor out of signup.ts into its own
src/auth/createPayment.ts brings signup.js back to 1.83KB and
createPayment.js lands at 1.25KB — both well under the cap, both
tree-shakable.

Phase 1 originally kept createPayment as a non-exported function inside
signup.ts to avoid auto-exposing it via the package's ./auth/* exports.
Phase 2 plans to make it public anyway (it's the shared primitive for
upgrade / credits / renewal). This fixup just lands the file shape early.

466/466 tests still green; behavior unchanged.
